Lau - corrección en motivos ausencia y sanciones

+++ Docs/07-Implementacion/Source/trunk/EDUAR/EDUAR_DataTransferObject/Entities/Package Usuarios/Docente.cs
@@ -7,15 +7,15 @@ namespace EDUAR_Entities
     [Serializable]
     public class Docente : Personal
     {
-        private Decimal _IdDocente;
-        public Decimal idDocente
+        private int _IdDocente;
+        public int idDocente
         {
             get { return _IdDocente; }
             set { _IdDocente = value; }
         }
 
-        private Decimal _IdDocenteTransaccional;
-        public Decimal idDocenteTransaccional
+        private int _IdDocenteTransaccional;
+        public int idDocenteTransaccional
         {
             get { return _IdDocenteTransaccional; }
             set { _IdDocenteTransaccional = value; }
